Fixing bad video player on Windows
On Windows native video player was used when user's screen was a touchscreen. I just deleted it and it solved the problem. So for everyone who had this issue, this is probably because you have a touchscreen.
Subtitles are now working fine and showing up (native video player only supports WEBVTT subtitles).