Fixes 2 bugs with fullscreen:
1. Properly toggle active class on $('.btn-os.fullscreen') when exiting
fullscreen from the maximize button. Before this commit the button class
won't be toggled and the button would show incorrect icons.

2. Fixes a bug where it got weird and unintuitive if the user wanted to
incorrectly unmaximize the window while being in fullscreen -- while the
window was already unmaximized (but in fullscreen). Now, if the user
clicks the maximize button while being in fullscreen, it not only exits
fullscreen but also unmaximizes it.