Don't play/pause toggle on player screen click
At least on OSX, its pretty common for players to allow you to drag the
window and it gets pretty annoying to pause the playback while doing so.
The click also gets in the way of the double click for quick fullscreen.

The solution is kinda hacky, as it monkey patches videojs.